\name{DSC_Static}
\alias{DSC_Static}
\title{DataStreamClusterer: Static}
\description{Creates a new DataStreamClusterer with fixed attributes}

\usage{
	DSC_Static(centers = NULL, microclusters = NULL, weights = NULL)
}

\arguments{
	\item{centers}{The centers of the clusters.}
	\item{microclusters}{The centers of the microclusters.}
	\item{weights}{The weights of the centers.}
}

\details{
This DSC is generated by \code{merge_DSC} and \code{prune_clusters}. This allows for the evaluation
of the results of the previous methods. Unlike other \code{DSC}s, however, clustering cannot be performed.
Once it has been generated it cannot change.
}

\value{
A list of class \code{DSC}, \code{DSC_R}, and \code{DSC_tNN}. The list
contains the following items:

	\item{description}{The name of the algorithm in the DSC object.}
	\item{RObj}{The underlying R object.}
}

\seealso{
	\code{\link{merge_DSC}}
	\code{\link{prune_clusters}}
}

\examples{
# 3 clusters with 5% noise
dsd <- DSD_Gaussian_Static(k=3, noise=0.05)

dsc <- DSC_tNN(r=0.05, lambda=.01, macro=FALSE)
cluster(dsc, dsd, 500)
dsc

# prune the data and collect the DSC_Static
dsc_static <- prune_clusters(dsc)

#plot the results
plot(get_microclusters(dsc_static))
}
