\name{get_centers}
\alias{get_centers}
\title{get_centers}
\description{Gets the centers of micro-clusters (if available) from a DSC object}

\usage{
	get_centers()
}

\arguments{
}

\details{
}

\value{
}

%%\references{ } TODO: need to reference the original paper

\seealso{
}

\examples{
}
