\name{DSC_ClusTree}
\alias{DSC_ClusTree}
\title{DataStreamClusterer: ClusTree}

\description{
Creates a new DataStreamClusterer with the ClusTree algorithm
}

\usage{
	DSC_ClusTree(horizon = 1000, maxHeight=8)
}

\arguments{
	\item{horizon}{Range of the (time) window.}
	\item{maxHeight}{The maximum height of the tree.}
}

\details{
The authors of ClusTree define it as a parameter free algorithm. Their
innovation automatically adjusts to the speed of the incoming points
generated by a data stream. ClusTree is able to detect concept drift and
outliers in the stream being operated upon.
}

\value{
A list of class \code{DSC}, \code{DSC_MOA}, and \code{DSC_ClusTree}. The list
contains the following items:

	\item{description}{The name of the algorithm in the DSC object.}
	\item{options}{The CLI params defined when creating the DSC object.}
	\item{javaObj}{The underlying Java object associated with DSC_MOA objects.}
}
\references{
Philipp Kranen, Ira Assent, Corinna Baldauf, and Thomas Seidl. 2009.
Self-Adaptive Anytime Stream Clustering. In Proceedings of the 2009 Ninth
IEEE International Conference on Data Mining (ICDM '09). IEEE Computer
Society, Washington, DC, USA, 249-258. DOI=10.1109/ICDM.2009.47
\url{http://dx.doi.org/10.1109/ICDM.2009.47}
}
\seealso{
	\code{\link{DSClusterer}},
	\code{\link{DSC_MOA}}
}
\examples{
dsd <- DSD_Gaussian_Static(k=3, noise=.05)

dsc <- DSC_ClusTree()
cluster(dsc, dsd, 500)
dsc

# plot micro-clusters
plot(get_points(dsd, 500), col="gray", main="ClusTree")
centers <- get_centers(dsc)
points(centers, col="red", cex=5)

# reclustering
cl <- kmeans(centers, centers=3)
points(cl$centers, col="blue", cex=15)

}
