\name{DSC_ClusTree}
\alias{DSC_ClusTree}
\title{DataStreamClusterer: ClusTree}

\description{
Creates a new DataStreamClusterer with the ClusTree algorithm
}

\usage{
	DSC_ClusTree(horizon = 1000, maxHeight = 8, negLambda = NULL)
}

\arguments{
	\item{horizon}{Range of the (time) window.}
	\item{maxHeight}{The maximum height of the tree.}
 	\item{negLambda}{number used to override computed lambda}
}

\details{
The authors of ClusTree define it as a parameter free algorithm. Their
innovation automatically adjusts to the speed of the incoming points
generated by a data stream. ClusTree is able to detect concept drift and
outliers in the stream being operated upon.
}

\value{
A list of class \code{DSC}, \code{DSC_MOA}, \code{DSC_Micro}, and \code{DSC_ClusTree}. The list
contains the following items:

	\item{description}{The name of the algorithm in the DSC object.}
	\item{options}{The CLI params defined when creating the DSC object.}
	\item{javaObj}{The underlying Java object associated with DSC_MOA objects.}
}
\references{
Philipp Kranen, Ira Assent, Corinna Baldauf, and Thomas Seidl. 2009.
Self-Adaptive Anytime Stream Clustering. In Proceedings of the 2009 Ninth
IEEE International Conference on Data Mining (ICDM '09). IEEE Computer
Society, Washington, DC, USA, 249-258. DOI=10.1109/ICDM.2009.47
\url{http://dx.doi.org/10.1109/ICDM.2009.47}
}
\seealso{
	\code{\link{DSClusterer}},
	\code{\link{DSC_MOA}}
}
\examples{
# 3 clusters with 5% noise
dsd <- DSD_Gaussian_Static(k=3, noise=0.05)

dsc <- DSC_ClusTree()
cluster(dsc, dsd, 1000)
dsc

# plot micro-clusters
plot(dsc,dsd)

# reclustering
kmeans <- DSC_KmeansW(3)
recluster(kmeans,dsc)
plot(kmeans,dsd,microclusters=TRUE)

}
