\name{DSD_moa}
\alias{DSD_moa}
\title{DataStreamData - MOA}
\description{
}
\usage{
DSD_MOA <- function(modelSeed=1, instanceSeed=1, numCluster=4L, avgRadius=0, density=0)
}
\details{
}
\value{
	\item{description}{The name of the class of the DSD object.}
	\item{options}{The CLI params defined when creating the DSD object.}
	\item{javaObj}{The underlying Java object associated with DSD_MOA objects.}
}
%%\references{} TODO: perhaps reference the randomrbfgenerator from the moa project?
%%\seealso{} TODO: reference the other DSD classes
%%\examples{} TODO: make an example
