\name{DSD_ReadStream}
\alias{DSD_ReadStream}
\title{DataStreamData: ReadStream (from an open connection or a file name)}
\description{

}
\usage{
DSD_ReadStream(x, delimiter = ",")
}
\arguments{
	\item{x}{An open connection, or a file path/URL to be opened as a connection.}
	\item{delimiter}{The character string that separates dimensions in data points in the stream.}
}
\details{
}
\value{
	\item{description}{The name of the class of the DSD object.}
	\item{con}{The connection where the data stream is being read from.}
	\item{delimiter}{The character string that separates dimensions in data points in the stream.}
}
%%\seealso{
%%}
\examples{
}
