\name{reset_stream}
\alias{reset_stream}
\title{reset_stream}
\description{Writes points to a connection from a DSD object.}

\usage{
	reset_stream()
}

\arguments{
}

\details{
}

\value{
}

%%\references{ } TODO: need to reference the original paper

\seealso{
}

\examples{
}
