\name{write_stream}
\alias{write_stream}
\title{write_stream}
\description{Writes points to a connection from a DSD object.}

\usage{
	write_stream()
}

\arguments{
}

\details{
}

\value{
}

%%\references{ } TODO: need to reference the original paper

\seealso{
}

\examples{
}
